api/translator.py
from dotenv import load_dotenv
load_dotenv()
import os
import six
from google.cloud import translate_v2 as translate
import json


def translate_text(text, target='en'):
    """Translates text into the target language.

    Target must be an ISO 639-1 language code.
    See https://g.co/cloud/translate/v2/translate-reference#supported_languages
    """
    json.load(os.environ.get("GOOGLE_APPLICATION_CREDENTIALS"))
    translate_client = translate.Client()

    if isinstance(text, six.binary_type):
        text = text.decode("utf-8")

    result = translate_client.translate(text, target_language=target)

    # print(u"Text: {}".format(result["input"]))
    # print(u"Translation: {}".format(result["translatedText"]))
    # print(u"Detected source language: {}".format(result["detectedSourceLanguage"]))

    return result["detectedSourceLanguage"]